Move event list items loading and displaying related code into new child fragment. #262
