+++ eventdiscovery-sdk/src/main/res/layout/schedjoules_fragment_event_list.xml
@@ -38,27 +38,10 @@
                 app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
 
-            <include
-                    android:id="@+id/schedjoules_event_list_include"
-                    layout="@layout/schedjoules_event_list"/>
-
-            <com.schedjoules.eventdiscovery.framework.widgets.StateSavingTextView
-                    android:id="@+id/schedjoules_event_list_background_message"
+            <FrameLayout
+                    android:id="@+id/schedjoules_event_list_list_holder"
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:layout_margin="@dimen/schedjoules_activity_horizontal_margin"
-                    android:paddingBottom="30dp"
-                    android:textSize="16sp"
-                    android:visibility="gone"
-                    android:gravity="center"
-                    android:layout_gravity="center"/>
-
-            <com.schedjoules.eventdiscovery.framework.widgets.AccentColoredProgressBar
-                    android:id="@+id/schedjoules_event_list_progress_bar"
-                    android:layout_gravity="center"
-                    android:visibility="gone"
-                    android:layout_width="@dimen/schedjoules_event_list_progress_bar"
-                    android:layout_height="@dimen/schedjoules_event_list_progress_bar"/>
+                    android:layout_height="match_parent"/>
 
             <View
                     android:layout_gravity="top"
